refactor(bootstrap): lift App.tsx's module-level bootstrap pipeline out of the render path (#141)
Split the 430-line module-level bootstrap pipeline that ran from App's
render path into three independently testable phases, leaving App with a
small composing function that owns the phase ordering:

- bootstrap/resolveWorkspace.ts — the six-way workspace resolution (pure
  async), moved verbatim.
- sync/keys/resolveWorkspaceEntry.ts — reunites the §6 gate's read-inputs
  half (pin + key-store + local row reads) with the decide half
  (decideWorkspaceEntry) it feeds. The cross-layer workspace-row read is
  injected so the module stays within sync/keys.
- bootstrap/workspaceBootstrap.ts — the bootstrap write phase plus the
  landing-runtime cache and URL→layout application; testable without
  rendering.

resolveInitialLayout is now just the composer: resolve → gate → write,
with the gate-before-writes ordering enforced by control flow rather than
comments. Adds a unit test for resolveWorkspaceEntry covering the key-store
read, its failure handling, and the lock-prompt name/canary enrichment.

App.tsx drops from 619 to 332 lines; behavior unchanged.
